refactor: reused types in global types folder
